+++ src/light_io.c
@@ -83,7 +83,7 @@ size_t light_io_write(light_file fd, const void* buf, size_t count)
 	return fd->fn_write(fd->context, buf, count);
 }
 
-int light_io_seek(light_file fd, long int offset, int origin)
+int64_t light_io_seek(light_file fd, int64_t offset, int origin)
 {
 	if (fd->fn_seek == NULL) {
 		return -1;
